Replace the housing system with the editor, directory and zoning work from PR 111
Takes the server-side housing implementation from
Open-Source-Free-Realms/Sanctuary#111 (raisingkaines) wholesale in place of our
own, keeping the pieces of ours it has no equivalent for.

PR 111 targets upstream's refactored database layout, which this fork never
adopted, so nothing is taken by path: the entities land in
Sanctuary.Database.Entities and the provider configs in Sanctuary.Database.SqLite
(whose folder differs from the assembly's Sqlite casing - checking out their
paths on Windows would collapse the two into one corrupt tree).

What arrives: the native decoration editor (place/move/rotate/scale/save, pick up
one or all, customization by fixture group and type), per-character ownership,
fixture-actor rehydration after zoning, the housing directory with ratings
(BaseRatingPacket + HouseVotes), empty-lot zones and the placement tables.

What is carried over from ours: DbHousePermission and the /createhouse, /gohouse
and /listhouses commands, whose raw SQL is rewritten for the new column names -
Houses.Id is ValueGeneratedNever now, so last_insert_rowid() no longer works.

Shared files were merged rather than replaced. Of note: op58 carries either a
player guid (friend teleport) or a point-of-interest id (atlas fast travel), and
the two implementations had each assumed only one, so the handler now resolves a
player first and falls back to the POI lookup. Player.TeleportToZone keeps its
explicit sky parameter, with null meaning "resolve the destination world's own",
which is what PR 111's lookup computes.

The migration drops and recreates the housing tables: the key types change from
ulong to int, SQLite cannot alter a column type, and the old NameId/IconId/
CustomName trio has no mapping into the new model. Existing houses are lost.

+++ src/Sanctuary.Gateway/Handlers/BaseHousingPacket/ClientHousingPacketToggleFloraAllowedHandler.cs
@@ -40,7 +40,7 @@ public static class ClientHousingPacketToggleFloraAllowedHandler
 
         var dbHouse = dbContext.Houses
             .Include(h => h.Fixtures)
-            .FirstOrDefault(h => h.Id == connection.Player.CurrentHouseGuid);
+            .FirstOrDefault(h => (ulong)h.Id == connection.Player.CurrentHouseGuid);
 
         if (dbHouse == null)
             return false;
